refactor: changed the structure of the tailwind config helper functions
