+++ utils/tailwind/figmaTokens.json
@@ -37,7 +37,6 @@
       "type": "letterSpacing"
     }
   },
-
   "boxShadow": {
     "shd-1": {
       "value": {
